My version of optimization/common.py has changed substantially from Luca's upstream version. To incorporate the changes from luca's upstream version, I am moving common -> common_new, and putting the upstream (old) common.py.
